@@ -1,5 +1,5 @@
 import * as React from "react";
-import { match, Route, Switch } from "react-router-dom";
+import { Route, Routes } from "react-router-dom";
 import type { ITeam } from "@shlack/types";
 import SelectedChannel from "./SelectedChannel";
 import TeamSidebar from "./TeamSidebar";
@@ -13,18 +13,14 @@ const Team: React.FunctionComponent<{ team: ITeam }> = ({ team }) => {
   return (
     <div className="flex-1 flex">
       <TeamSidebar team={team} />
-      <Switch>
-        <Route exact path={`/team/${team.id}`}>
+      <Routes>
+        <Route path={`/team/${team.id}`}>
           <h3>Please select a channel</h3>
         </Route>
-        <Route
-          exact
-          path={`/team/${team.id}/channel/:channelId`}
-          children={({ match }: { match: match<{ channelId: string }> }) => (
-            <SelectedChannel match={match} channels={channels} />
-          )}
-        />
-      </Switch>
+        <Route path={`/team/${team.id}/channel/:channelId`}>
+          <SelectedChannel channels={channels} />
+        </Route>
+      </Routes>
     </div>
   );
 };
