@@ -1,9 +1,9 @@
 import * as React from "react";
 import {
-  BrowserRouter as Router,
-  match,
+  createBrowserRouter,
+  createRoutesFromElements,
   Route,
-  Switch,
+  RouterProvider,
 } from "react-router-dom";
 import { getAllTeams } from "@shlack/data";
 import { ITeam } from "@shlack/types";
@@ -12,40 +12,45 @@ import Loading from "./components/Loading";
 import SelectedTeam from "./components/SelectedTeam";
 import TeamSelector from "./components/TeamSelector";
 
-const { useState } = React;
+const createRouter = (teams: ITeam[]) => {
+  return createBrowserRouter(
+    createRoutesFromElements(
+      <>
+        <Route path="/">
+          <section className="m-12 text-xl">
+            <h3>Please select a team</h3>
+          </section>
+        </Route>
+        <Route path="/team">
+          <section className="m-12 text-xl">
+            <h3>Please select a team</h3>
+          </section>
+        </Route>
+        <Route path="/team/:teamId">
+          <SelectedTeam teams={teams} />
+        </Route>
+      </>
+    )
+  );
+};
 
 const App: React.FunctionComponent = () => {
-  const [teams, setTeams] = useState<ITeam[]>();
+  const [teams, setTeams] = React.useState<ITeam[]>();
 
   useAsyncDataEffect(() => getAllTeams(), {
     setter: setTeams,
     stateName: "teams",
   });
+
   if (!teams) return <Loading message="Loading teams" />;
+
+  const router = createRouter(teams);
+
   return (
-    <Router>
-      <div className="flex flex-col sm:flex-row w-full h-full">
-        <TeamSelector teams={teams} />
-        <Switch>
-          <Route exact path="/">
-            <section className="m-12 text-xl">
-              <h3>Please select a team</h3>
-            </section>
-          </Route>
-          <Route exact path="/team">
-            <section className="m-12 text-xl">
-              <h3>Please select a team</h3>
-            </section>
-          </Route>
-          <Route
-            path="/team/:teamId"
-            children={({ match }: { match: match<{ teamId: string }> }) => (
-              <SelectedTeam match={match} teams={teams} />
-            )}
-          />
-        </Switch>
-      </div>
-    </Router>
+    <div className="flex flex-col sm:flex-row w-full h-full">
+      <TeamSelector teams={teams} />
+      <RouterProvider router={router} />
+    </div>
   );
 };
 export default App;
