@@ -1,9 +1,9 @@
 import * as React from "react";
 
-const Loading: React.FunctionComponent<{ message: string }> = ({
-  message = "Loading...",
-  children,
-}) => (
+const Loading: React.FunctionComponent<{
+  children?: React.ReactNode;
+  message: string;
+}> = ({ message = "Loading...", children }) => (
   <h1>
     {message}...
     {children}
