@@ -1,26 +1,19 @@
 import * as React from "react";
-import { match } from "react-router";
 import { IChannel } from "@shlack/types";
 import Channel from "./Channel";
+import { useParams } from "react-router-dom";
 
 const SelectedChannel: React.FunctionComponent<{
-  match: match<{ channelId: string }>;
   channels: IChannel[];
-}> = ({ match, channels }) => {
-  if (!channels) throw new Error("no channels");
-  if (!match) throw new Error("no match");
+}> = ({ channels }) => {
+  const { channelId } = useParams();
 
-  const { params } = match;
-  if (!match) return <p>No match params</p>;
-  const { channelId: selectedChannelId } = params;
-  if (!selectedChannelId) return <p>Invalid channelId</p>;
-  const selectedChannel = channels.find(
-    (c: IChannel) => c.id === selectedChannelId
-  );
+  if (!channelId) return <p>Invalid channelId</p>;
+  const selectedChannel = channels.find((c: IChannel) => c.id === channelId);
   if (!selectedChannel)
     return (
       <div>
-        <p>Could not find channel with id {selectedChannelId}</p>
+        <p>Could not find channel with id {channelId}</p>
         <pre>{JSON.stringify(channels, null, "  ")}</pre>
       </div>
     );
