@@ -1,12 +1,13 @@
 import "regenerator-runtime/runtime";
+import { createRoot } from "react-dom/client";
 import * as React from "react";
-import { render } from "react-dom";
 import App from "./App";
 
 function initializeReactApp() {
   const appContainer = document.getElementById("appContainer");
   if (!appContainer) throw new Error("No #appContainer found in DOM");
-  render(React.createElement(App), appContainer);
+  const root = createRoot(appContainer);
+  root.render(<App />);
 }
 
 initializeReactApp();
